+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History of versions #
 
+* Version 3.13.0 (2026-09-13)
+  * Allow to set the default rows and columns size
+  * Fixed macro editing bug
+* Version 3.12.0 (2026-09-11)
+  * Fix dock icon when opening a new window
 * Version 3.11.0 (2026-07-25)
   * Create the xdg-terminals.list if it doesn't exist
 * Version 3.10.0 (2026-07-19)